Remove PlanningSpikeService as it is no longer needed - function needed it changed to async
