+++ server/planning/events/events_spike.py
@@ -264,7 +264,7 @@ async def process_spike_event(updates: dict[str, Any], original: dict[str, Any])
 
     # Clean updates before persisting change
     spiked_items = updates.pop("_spiked_items", [])
-    remove_autosave_on_spike(original)
+    await remove_autosave_on_spike(original)
     updates.pop("update_method", None)
     updates.pop("skip_on_update", None)
